move wagmi stuff into core
