+++ src/core/createWagmiConfig.ts
@@ -1,7 +1,7 @@
 import { http, cookieStorage, createConfig, createStorage } from 'wagmi';
 import { base, baseSepolia } from 'wagmi/chains';
 import { coinbaseWallet } from 'wagmi/connectors';
-import type { CreateWagmiConfigParams } from '../../types';
+import type { CreateWagmiConfigParams } from '../types';
 
 // createWagmiConfig returns a WagmiConfig (https://wagmi.sh/react/api/createConfig) using OnchainKit provided settings.
 // This function should only be used if the user does not provide WagmiProvider as a parent in the React context.
